refactor: extract runs/workflows, assistant, and feature-requests route groups (batch 2)
Part of epic #159. Reduces server.py by 1,138 lines (6,661 → 5,523).

Extracted route groups:
- `routers/runs_workflows.py`: GET/POST /api/runs, /api/runs/enriched,
  /api/runs/{repo}, /api/scheduled-workflows, /api/workflows/list,
  /api/workflows/dispatch (393 lines)
- `routers/assistant.py`: POST/GET /api/assistant/chat, /api/assistant/tool/execute,
  /api/assistant/audit-history, /api/assistant/propose-action,
  /api/assistant/execute-action (369 lines)
- `routers/feature_requests.py`: GET/POST /api/feature-requests,
  /api/feature-requests/templates, /api/feature-requests/dispatch,
  GET/PUT /api/settings/prompt-notes (287 lines)

All three routers use existing utility modules (cache_utils, dashboard_config,
gh_utils, proxy_utils, security, system_utils) and are registered via
app.include_router() in server.py.
