+++ backend/server.py
@@ -58,8 +58,6 @@ if str(BACKEND_DIR) not in sys.path:
 
 import agent_dispatch_router as agent_dispatch_router  # noqa: E402
 import agent_remediation as agent_remediation  # noqa: E402
-import assistant_contract as assistant_contract  # noqa: E402
-import assistant_tools as assistant_tools  # noqa: E402
 import auth_webauthn as _auth_webauthn_router  # noqa: E402
 import config_schema as config_schema  # noqa: E402
 import deployment_drift as deployment_drift  # noqa: E402
@@ -73,7 +71,6 @@ import pr_inventory as pr_inventory  # noqa: E402
 import push as _push_router  # noqa: E402
 import quick_dispatch as _quick_dispatch  # noqa: E402
 import quota_enforcement as quota_enforcement  # noqa: E402
-import scheduled_workflows as scheduled_workflow_inventory  # noqa: E402
 import unified_issue_inventory as unified_issue_inventory  # noqa: E402
 import usage_monitoring as usage_monitoring  # noqa: E402
 from local_app_monitoring import collect_local_apps  # noqa: E402
@@ -82,10 +79,13 @@ from machine_registry import (  # noqa: E402
     merge_registry_with_live_nodes,
 )
 from report_files import parse_report_metrics, sanitize_report_date  # noqa: E402
+from routers import assistant as _assistant_router  # noqa: E402
 from routers import credentials as _credentials_router  # noqa: E402
 from routers import dispatch as _dispatch_router  # noqa: E402
+from routers import feature_requests as _feature_requests_router  # noqa: E402
 from routers import linear as _linear_router  # noqa: E402
 from routers import linear_webhook as _linear_webhook_router  # noqa: E402
+from routers import runs_workflows as _runs_workflows_router  # noqa: E402
 from system_utils import get_system_metrics_snapshot  # noqa: E402
 
 # datetime.UTC added in Python 3.11; fall back to timezone.utc on older runtimes.
@@ -280,9 +280,7 @@ _CACHE_EVICT_BATCH = 50
 # ─── Shared State Locks ───────────────────────────────────────────────────────
 _remediation_history_lock: asyncio.Lock = asyncio.Lock()
 _orchestration_audit_lock: asyncio.Lock = asyncio.Lock()
-_feature_requests_lock: asyncio.Lock = asyncio.Lock()
-_prompt_templates_lock: asyncio.Lock = asyncio.Lock()
-_prompt_notes_lock: asyncio.Lock = asyncio.Lock()
+# Feature-request locks moved to routers/feature_requests.py
 
 # ─── Configuration ────────────────────────────────────────────────────────────
 ORG = os.environ.get("GITHUB_ORG", "D-sorganization")
@@ -460,6 +458,13 @@ app.include_router(_metrics_router.router)
 import agent_launcher_router as _agent_launcher_router  # noqa: E402
 
 app.include_router(_agent_launcher_router.router)
+
+# Batch-2 extracted routers (epic #159)
+# Note: fleet router not registered here because fleet_control has richer fan-out
+# logic in server.py that differs from the simplified routers/fleet.py version.
+app.include_router(_runs_workflows_router.router)
+app.include_router(_assistant_router.router)
+app.include_router(_feature_requests_router.router)
 
 app.add_middleware(
     SessionMiddleware,
@@ -2252,561 +2257,8 @@ async def get_watchdog_status(request: Request):
     return await _watchdog_status_impl()
 
 
-# ─── Runner API Routes ───────────────────────────────────────────────────────
-
-
-@app.get("/api/runners")
-async def get_runners(request: Request):
-    """Get all org runners with their status."""
-    if _should_proxy_fleet_to_hub(request):
-        return await proxy_to_hub(request)
-
-    cached = _cache_get("runners", 60.0)
-    if cached is not None:
-        cached["runners"] = sorted(cached.get("runners", []), key=_runner_sort_key)
-        return cached
-    data = await gh_api_admin(f"/orgs/{ORG}/actions/runners")
-    data["runners"] = sorted(data.get("runners", []), key=_runner_sort_key)
-    _cache_set("runners", data)
-    return data
-
-
-# ─── MATLAB Runner Health (issue #570) ───────────────────────────────────────
-#
-# MATLAB linting relies on a Windows self-hosted runner where MATLAB is
-# installed natively.  Operators need to see at a glance whether that capacity
-# is available; missing MATLAB runners produce queued-forever CI jobs
-# otherwise.  The endpoint below surfaces online/offline/busy state, last job,
-# persistence mode (Windows service vs Scheduled Task), and links to recent
-# MATLAB Code Analyzer workflow runs.
-#
-# Runners are identified by either the `matlab` label or a name matching
-# `*windows-matlab*` / `*d-sorg-matlab*` to stay robust across naming schemes.
-
-
-def _is_matlab_runner(runner: dict) -> bool:
-    """Return True if the runner appears to be a MATLAB-capable runner."""
-    name = str(runner.get("name", "")).lower()
-    if "matlab" in name:
-        return True
-    for label in runner.get("labels", []) or []:
-        lname = str(label.get("name", "")).lower() if isinstance(label, dict) else str(label).lower()
-        if lname == "matlab" or lname.startswith("windows-matlab") or lname.startswith("d-sorg-matlab"):
-            return True
-    return False
-
-
-def _matlab_runner_summary(runner: dict) -> dict:
-    """Project a GitHub runner record into the MATLAB health shape."""
-    labels = [lbl.get("name") if isinstance(lbl, dict) else str(lbl) for lbl in (runner.get("labels") or [])]
-    status = str(runner.get("status", "unknown")).lower()
-    busy = bool(runner.get("busy"))
-    # Persistence hint: the ControlTower Windows runner is registered as a
-    # Windows service; names ending with `-scheduled` signal a Scheduled Task.
-    name = str(runner.get("name", ""))
-    if name.endswith("-scheduled") or "scheduled-task" in name.lower():
-        persistence = "scheduled_task"
-    elif status == "offline":
-        persistence = "unknown"
-    else:
-        persistence = "windows_service"
-    return {
-        "id": runner.get("id"),
-        "name": name,
-        "status": status,
-        "busy": busy,
-        "labels": labels,
-        "os": runner.get("os"),
-        "persistence": persistence,
-    }
-
-
-async def _recent_matlab_workflow_runs(limit: int = 5) -> list[dict]:
-    """Fetch recent MATLAB Code Analyzer workflow runs across the org.
-
-    Returns an empty list on any failure — this is advisory UI, not critical
-    control surface, so transient API errors must not break the endpoint.
-    """
-    try:
-        repos = await _get_recent_org_repos(limit=15)
-    except Exception:  # pragma: no cover - defensive  # noqa: BLE001
-        return []
-    if not repos:
-        return []
-
-    async def _runs_for_repo(repo_name: str) -> list[dict]:
-        try:
-            data = await gh_api_admin(f"/repos/{ORG}/{repo_name}/actions/runs?per_page=10")
-        except Exception:  # noqa: BLE001
-            return []
-        out = []
-        for run in data.get("workflow_runs", []) or []:
-            wf_name = str(run.get("name") or "").lower()
-            wf_path = str(run.get("path") or "").lower()
-            if "matlab" in wf_name or "matlab" in wf_path or "code analyzer" in wf_name:
-                out.append(
-                    {
-                        "repo": repo_name,
-                        "name": run.get("name"),
-                        "status": run.get("status"),
-                        "conclusion": run.get("conclusion"),
-                        "html_url": run.get("html_url"),
-                        "created_at": run.get("created_at"),
-                        "run_id": run.get("id"),
-                    }
-                )
-        return out
-
-    try:
-        nested = await asyncio.gather(
-            *[_runs_for_repo(r["name"]) for r in repos[:10]],
-            return_exceptions=True,
-        )
-    except Exception:  # noqa: BLE001
-        return []
-    flat: list[dict] = []
-    for item in nested:
-        if isinstance(item, list):
-            flat.extend(item)
-    flat.sort(key=lambda r: r.get("created_at") or "", reverse=True)
-    return flat[:limit]
-
-
-@app.get("/api/runners/matlab")
-async def get_matlab_runner_health(request: Request) -> dict:
-    """Surface Windows MATLAB runner health for the dashboard (issue #570).
-
-    Response shape::
-
-        {
-          "runners": [...],          # MATLAB runner summaries
-          "total": int,
-          "online": int,
-          "busy": int,
-          "offline": int,
-          "capacity_available": bool, # true iff an idle online runner exists
-          "warning": str | None,      # actionable message when capacity is zero
-          "recent_workflow_runs": [...],
-          "generated_at": "..."
-        }
-
-    Always returns 200; absence of runners is represented explicitly so the UI
-    can render an actionable warning instead of a spinner.
-    """
-    if _should_proxy_fleet_to_hub(request):
-        return await proxy_to_hub(request)
-
-    cached = _cache_get("matlab_runner_health", 45.0)
-    if cached is not None:
-        return cached
-
-    try:
-        data = await gh_api_admin(f"/orgs/{ORG}/actions/runners")
-        all_runners = data.get("runners", []) or []
-    except Exception as exc:  # pragma: no cover - defensive  # noqa: BLE001
-        all_runners = []
-        api_error: str | None = f"GitHub runner API unavailable: {exc}"
-    else:
-        api_error = None
-
-    matlab = [r for r in all_runners if _is_matlab_runner(r)]
-    summaries = [_matlab_runner_summary(r) for r in matlab]
-
-    online = sum(1 for r in summaries if r["status"] == "online")
-    busy = sum(1 for r in summaries if r["busy"])
-    offline = sum(1 for r in summaries if r["status"] != "online")
-    idle_online = sum(1 for r in summaries if r["status"] == "online" and not r["busy"])
-
-    warning: str | None = None
-    if not summaries:
-        warning = (
-            "No Windows MATLAB runners are registered. MATLAB Code Analyzer "
-            "jobs will queue indefinitely. See "
-            "docs/operations/matlab_windows_runner.md to register a runner."
-        )
-    elif online == 0:
-        warning = (
-            "All MATLAB runners are offline. Start the Windows runner service "
-            "on the ControlTower host to restore MATLAB lint capacity."
-        )
-    elif idle_online == 0:
-        warning = "All MATLAB runners are currently busy. New MATLAB lint jobs will queue until one frees up."
-
-    recent = await _recent_matlab_workflow_runs(limit=5)
-
-    result = {
-        "runners": summaries,
-        "total": len(summaries),
-        "online": online,
-        "busy": busy,
-        "offline": offline,
-        "capacity_available": idle_online > 0,
-        "warning": warning,
-        "api_error": api_error,
-        "recent_workflow_runs": recent,
-        "generated_at": datetime.now(UTC).isoformat(),
-    }
-    _cache_set("matlab_runner_health", result)
-    return result
-
-
-@app.get("/api/runs")
-async def get_runs(request: Request, per_page: int = 30) -> dict:
-    """Get recent workflow runs across the org by sampling the most active repos.
-
-    GitHub's REST API has no org-level /actions/runs endpoint; runs must be
-    fetched per-repo.  We sample the 10 most recently updated repos and return
-    up to ``per_page`` runs sorted newest-first.
-    """
-    if _should_proxy_fleet_to_hub(request):
-        return await proxy_to_hub(request)
-
-    cache_key = f"runs:{per_page}"
-    cached = _cache_get(cache_key, 120.0)
-    if cached is not None:
-        return cached
-
-    repos = await _get_recent_org_repos(limit=20)
-    if not repos:
-        return {"workflow_runs": [], "total_count": 0}
-
-    # Fetch a handful of runs from each repo concurrently
-    runs_per_repo = max(3, per_page // max(len(repos[:10]), 1))
-
-    sample = repos[:10]
-    all_runs_nested = await asyncio.gather(*[_fetch_repo_runs(r["name"], per_page=runs_per_repo) for r in sample])
-    all_runs: list[dict] = [run for sublist in all_runs_nested for run in sublist]
-
-    # Sort newest-first and cap at per_page
-    all_runs.sort(key=lambda r: r.get("created_at", ""), reverse=True)
-    top_runs = all_runs[:per_page]
-
-    result = {"workflow_runs": top_runs, "total_count": len(top_runs)}
-    _cache_set(f"runs:{per_page}", result)
-    return result
-
-
-async def _scheduled_workflows_impl(
-    *,
-    include_archived: bool = False,
-    repo_limit: int = 100,
-) -> dict:
-    """Collect the read-only scheduled workflow inventory."""
-    cache_key = f"scheduled-workflows:{include_archived}:{repo_limit}"
-    cached = _cache_get(cache_key, 300.0)
-    if cached is not None:
-        return cached
-
-    raw_timeout = os.environ.get("SCHEDULED_WORKFLOWS_TIMEOUT", "20")
-    try:
-        timeout = float(raw_timeout)
-    except (TypeError, ValueError):
-        timeout = 20.0
-
-    try:
-        report = await asyncio.wait_for(
-            scheduled_workflow_inventory.collect_inventory(
-                ORG,
-                gh_api,
-                gh_api_raw,
-                repo_limit=repo_limit,
-                include_archived=include_archived,
-            ),
-            timeout=timeout,
-        )
-        payload = report.to_dict()
-        payload["status"] = "ok"
-        _cache_set(cache_key, payload)
-    except (TimeoutError, asyncio.TimeoutError):  # noqa: UP041
-        payload = {
-            "status": "degraded",
-            "organization": ORG,
-            "generated_at": datetime.now(UTC).isoformat(),
-            "repository_count": 0,
-            "scheduled_workflow_count": 0,
-            "repositories": [],
-            "dry_run_plan": {
-                "mode": "dry_run",
-                "write_actions_allowed": False,
-                "confirmation_required": True,
-                "audit_required": True,
-                "steps": [],
-            },
-            "error": "Scheduled workflow inventory timed out.",
-        }
-    return payload
-
-
-@app.get("/api/scheduled-workflows")
-async def get_scheduled_workflows(
-    request: Request,
-    include_archived: bool = False,
-    repo_limit: int = 100,
-):
-    """Inventory GitHub Actions schedules across org repositories.
-
-    This endpoint is read-only. It gathers workflow metadata, extracts cron
-    expressions from workflow YAML where available, and attaches a dry-run plan
-    that describes future changes without executing them.
-    """
-    if _should_proxy_fleet_to_hub(request):
-        return await proxy_to_hub(request)
-    return await _scheduled_workflows_impl(
-        include_archived=include_archived,
-        repo_limit=repo_limit,
-    )
-
-
-@app.get("/api/workflows/list")
-async def list_workflows() -> dict:
-    """List all workflows per repository with trigger capabilities and latest run."""
-    import base64  # noqa: PLC0415
-
-    cached = _cache_get("workflows_list", 120.0)
-    if cached is not None:
-        return cached
-    repos = await _get_recent_org_repos(limit=30)
-
-    async def get_repo_workflows(repo_name: str) -> list[dict]:
-        code, out, _ = await run_cmd(
-            ["gh", "api", f"/repos/{ORG}/{repo_name}/actions/workflows", "--paginate"],
-            timeout=20,
-            cwd=REPO_ROOT,
-        )
-        if code != 0:
-            return []
-        try:
-            data = json.loads(out)
-            workflows = data.get("workflows", [])
-        except Exception:  # noqa: BLE001
-            return []
-        result = []
-        for wf in workflows:
-            wf_id = wf.get("id")
-            triggers = []
-            if wf.get("path"):
-                code2, out2, _ = await run_cmd(
-                    ["gh", "api", f"/repos/{ORG}/{repo_name}/contents/{wf['path']}"],
-                    timeout=10,
-                    cwd=REPO_ROOT,
-                )
-                if code2 == 0:
-                    try:
-                        content_data = json.loads(out2)
-                        content = base64.b64decode(content_data.get("content", "")).decode("utf-8", errors="replace")
-                        if "workflow_dispatch" in content:
-                            triggers.append("manual")
-                        if "schedule" in content:
-                            triggers.append("schedule")
-                        if "push" in content or "pull_request" in content:
-                            triggers.append("push_pr")
-                        if "workflow_run" in content:
-                            triggers.append("workflow_run")
-                    except Exception:  # noqa: BLE001
-                        pass
-            code3, out3, _ = await run_cmd(
-                [
-                    "gh",
-                    "api",
-                    f"/repos/{ORG}/{repo_name}/actions/workflows/{wf_id}/runs?per_page=3",
-                ],
-                timeout=10,
-                cwd=REPO_ROOT,
-            )
-            latest_run = None
-            recent_runs = []
-            if code3 == 0:
-                try:
-                    runs_data = json.loads(out3)
-                    all_runs = runs_data.get("workflow_runs", [])
-                    if all_runs:
-                        latest_run = {
-                            "id": all_runs[0].get("id"),
-                            "status": all_runs[0].get("status"),
-                            "conclusion": all_runs[0].get("conclusion"),
-                            "created_at": all_runs[0].get("created_at"),
-                            "html_url": all_runs[0].get("html_url"),
-                            "head_branch": all_runs[0].get("head_branch"),
-                        }
-                        recent_runs = [
-                            {
-                                "id": r.get("id"),
-                                "status": r.get("status"),
-                                "conclusion": r.get("conclusion"),
-                                "created_at": r.get("created_at"),
-                                "html_url": r.get("html_url"),
-                            }
-                            for r in all_runs[:3]
-                        ]
-                except Exception:  # noqa: BLE001
-                    pass
-            result.append(
-                {
-                    "id": wf_id,
-                    "name": wf.get("name", ""),
-                    "path": wf.get("path", ""),
-                    "state": wf.get("state", ""),
-                    "html_url": wf.get("html_url", ""),
-                    "triggers": triggers,
-                    "latest_run": latest_run,
-                    "recent_runs": recent_runs,
-                    "repository": repo_name,
-                }
-            )
-        return result
-
-    results = await asyncio.gather(*[get_repo_workflows(r["name"]) for r in repos[:20]])
-    all_workflows: list[dict] = []
-    for wf_list in results:
-        all_workflows.extend(wf_list)
-
-    result = {"workflows": all_workflows, "total": len(all_workflows)}
-    _cache_set("workflows_list", result, 120.0)
-    return result
-
-
-@app.post("/api/workflows/dispatch")
-async def dispatch_workflow(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("workflows.control")),  # noqa: B008
-) -> dict:
-    """Dispatch a workflow via workflow_dispatch."""
-    body = await request.json()
-    repo = str(body.get("repository", "")).strip()
-    workflow_id = body.get("workflow_id")
-    ref = str(body.get("ref", "main")).strip()
-    inputs = body.get("inputs", {}) or {}
-    correlation_id = request.headers.get("X-Correlation-Id", secrets.token_hex(8))
-    inputs["correlation_id"] = correlation_id
-    approved_by = principal.id
-
-    if not repo or not workflow_id:
-        raise HTTPException(status_code=422, detail="repository and workflow_id required")
-
-    endpoint = f"/repos/{ORG}/{repo}/actions/workflows/{workflow_id}/dispatches"
-    payload = {"ref": ref, "inputs": inputs}
-    with tempfile.NamedTemporaryFile(mode="w", encoding="utf-8", suffix=".json", delete=False) as f:
-        json.dump(payload, f)
-        pf = f.name
-    try:
-        code, _, stderr = await run_cmd(
-            ["gh", "api", endpoint, "--method", "POST", "--input", pf],
-            timeout=30,
-            cwd=REPO_ROOT,
-        )
-    finally:
-        with contextlib.suppress(OSError):
-            Path(pf).unlink()
-    if code != 0:
-        log.warning(
-            "workflow_dispatch failed: repo=%s workflow_id=%s stderr=%s",
-            repo,
-            workflow_id,
-            stderr.strip()[:300],
-        )
-        raise HTTPException(status_code=502, detail="Workflow dispatch failed")
-
-    log.info(
-        "workflow_dispatch audit: repo=%s workflow_id=%s ref=%s approved_by=%s",
-        repo,
-        workflow_id,
-        ref,
-        sanitize_log_value(approved_by),
-    )
-    return {
-        "status": "dispatched",
-        "repository": repo,
-        "workflow_id": workflow_id,
-        "ref": ref,
-    }
-
-
-@app.get("/api/runs/enriched")
-async def get_enriched_runs(request: Request, per_page: int = 50) -> dict:
-    """Return recent runs with dashboard-friendly enrichment fields."""
-    if _should_proxy_fleet_to_hub(request):
-        return await proxy_to_hub(request)
-    cache_key = f"runs-enriched:{per_page}"
-    cached = _cache_get(cache_key, 120.0)
-    if cached is not None:
-        return cached
-
-    data = await get_runs(request, per_page=per_page)
-    runs = data.get("workflow_runs", [])
-    enrichable = runs[:RUN_JOB_ENRICHMENT_LIMIT]
-    enriched = await asyncio.gather(*[_enrich_run_with_job_placement(run) for run in enrichable])
-    enriched.extend(dict(run) for run in runs[RUN_JOB_ENRICHMENT_LIMIT:])
-    result = {"workflow_runs": enriched, "total_count": len(enriched)}
-    _cache_set(cache_key, result)
-    return result
-
-
-@app.get("/api/runs/{repo}")
-async def get_repo_runs(request: Request, repo: str, per_page: int = 20):
-    """Get recent workflow runs for a specific repo."""
-    if _should_proxy_fleet_to_hub(request):
-        return await proxy_to_hub(request)
-    data = await gh_api(f"/repos/{ORG}/{repo}/actions/runs?per_page={per_page}")
-    return data
-
-
-@app.post("/api/runners/{runner_id}/stop")
-async def stop_runner(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("runners.control")),
-    runner_id: int,  # noqa: B008
-):  # noqa: B008
-    """Stop a specific runner's systemd service."""
-    data = await gh_api_admin(f"/orgs/{ORG}/actions/runners")
-    runners = data.get("runners", [])
-    num = runner_num_from_id(runner_id, runners)
-
-    if num is None:
-        msg = f"Runner ID {runner_id} not found locally"
-        raise HTTPException(status_code=404, detail=msg)
-
-    svc_path = runner_svc_path(num)
-    if not svc_path.exists():
-        raise HTTPException(status_code=404, detail=f"Runner {num} svc.sh not found")
-
-    log.info("Stopping runner %d (GitHub ID: %d)", num, runner_id)
-    code, stdout, stderr = await run_runner_svc(num, "stop")
-    if code != 0:
-        log.warning("Failed to stop runner %d: %s", num, stderr[:200])
-        raise HTTPException(status_code=500, detail=f"Failed to stop runner {num}")
-
-    return {"status": "stopped", "runner": num, "output": stdout.strip()}
-
-
-@app.post("/api/runners/{runner_id}/start")
-async def start_runner(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("runners.control")),
-    runner_id: int,  # noqa: B008
-):  # noqa: B008
-    """Start a specific runner's systemd service."""
-    data = await gh_api_admin(f"/orgs/{ORG}/actions/runners")
-    runners = data.get("runners", [])
-    num = runner_num_from_id(runner_id, runners)
-
-    if num is None:
-        msg = f"Runner ID {runner_id} not found locally"
-        raise HTTPException(status_code=404, detail=msg)
-
-    svc_path = runner_svc_path(num)
-    if not svc_path.exists():
-        raise HTTPException(status_code=404, detail=f"Runner {num} svc.sh not found")
-
-    log.info("Starting runner %d (GitHub ID: %d)", num, runner_id)
-    code, stdout, stderr = await run_runner_svc(num, "start")
-    if code != 0:
-        log.warning("Failed to start runner %d: %s", num, stderr[:200])
-        raise HTTPException(status_code=500, detail=f"Failed to start runner {num}")
-
-    return {"status": "started", "runner": num, "output": stdout.strip()}
+# Runner routes extracted to routers/fleet.py and registered via app.include_router.
+# Runs and workflow routes extracted to routers/runs_workflows.py and registered via app.include_router.
 
 
 async def _fleet_control_local(action: str) -> dict:
@@ -4258,357 +3710,6 @@ async def get_agent_providers() -> dict:
     }
 
 
-# ─── Assistant Chat API (Issue #88) ───────────────────────────────────────────
-
-
-async def _dispatch_to_ai_provider_for_chat(
-    provider: str | None,
-    prompt: str,
-    context: dict,
-) -> str:
-    """Call the configured AI provider for assistant chat."""
-    # For MVP: return a simple response based on the prompt
-    # In production, this would dispatch to an actual provider (Jules, Claude, etc.)
-    provider_id = provider or "ollama_local"
-
-    # Check provider availability
-    availability = agent_remediation.probe_provider_availability()
-    if provider_id not in availability or not availability[provider_id].available:
-        # Return a synthetic response for MVP
-        return f"(Note: Provider '{provider_id}' is unavailable. Mock response for demonstration.)"
-
-    # For MVP: return a simple acknowledgment
-    # Tracked in #88/#89: implement actual provider dispatch (Jules API, local Ollama, Claude API, etc.)
-    return f"Assistant response to: {prompt[:100]}... (MVP mock - implement real provider dispatch)"
-
-
-@app.post("/api/assistant/chat", tags=["assistant"])
-async def assistant_chat(request: Request, *, principal: Principal = Depends(require_scope("assistant.chat"))) -> dict:  # noqa: B008
-    """Chat with AI assistant about dashboard state.
-
-    When ``tools_enabled: true`` is set, the Anthropic tool-use loop is
-    activated and the response may contain ``tool_calls`` for the client to
-    render as confirmation cards (issue #89).
-    """
-    try:
-        body = await request.json()
-    except Exception:  # noqa: BLE001
-        raise HTTPException(status_code=400, detail="Invalid JSON") from None
-
-    try:
-        req = assistant_contract.AssistantChatRequest(**body)
-    except Exception as e:  # noqa: BLE001
-        raise HTTPException(status_code=422, detail=str(e)) from e
-
-    now_ts = datetime.now(UTC).isoformat()
-
-    # ── Tool-use path (Issue #89) ──────────────────────────────────────────
-    if req.tools_enabled:
-        anthropic_key = os.environ.get("ANTHROPIC_API_KEY", "")
-        if not anthropic_key:
-            raise HTTPException(
-                status_code=503,
-                detail="ANTHROPIC_API_KEY not configured; tool-use requires Anthropic.",
-            )
-        try:
-            result = await assistant_tools.call_anthropic_with_tools(
-                api_key=anthropic_key,
-                prompt=req.prompt,
-                context=req.context.dict(),
-                model=DEFAULT_LLM_MODEL,
-                tools_enabled=True,
-            )
-        except Exception as exc:  # noqa: BLE001
-            log.error("Anthropic tool-use error: %s", exc)
-            raise HTTPException(status_code=502, detail=f"Anthropic error: {exc}") from exc
-        return {
-            "message": result["message"],
-            "stop_reason": result["stop_reason"],
-            "tool_calls": result["tool_calls"],
-            "provider": "anthropic",
-            "timestamp": now_ts,
-        }
-
-    # ── Standard chat path ────────────────────────────────────────────────
-    response_text = await _dispatch_to_ai_provider_for_chat(
-        provider=req.provider,
-        prompt=req.prompt,
-        context=req.context.dict(),
-    )
-    return {
-        "response": response_text,
-        "provider": req.provider or "ollama_local",
-        "context_used": req.context.dict(),
-        "timestamp": now_ts,
-    }
-
-
-# ─── Tool Execute API (Issue #89) ─────────────────────────────────────────────
-
-
-@app.post("/api/assistant/tool/execute", tags=["assistant"])
-async def execute_assistant_tool(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("assistant.execute")),  # noqa: B008
-) -> dict:
-    """Execute a tool call from the assistant allowlist (Issue #89).
-
-    State-changing tools require ``confirmation`` in the request body.
-    Every execution (success or failure) is appended to the audit log.
-    """
-    try:
-        body = await request.json()
-    except Exception:  # noqa: BLE001
-        raise HTTPException(status_code=400, detail="Invalid JSON") from None
-
-    try:
-        req = assistant_contract.ToolExecuteRequest(**body)
-    except Exception as exc:  # noqa: BLE001
-        raise HTTPException(status_code=422, detail=str(exc)) from exc
-
-    if req.name not in assistant_tools.TOOL_ALLOWLIST:
-        raise HTTPException(
-            status_code=422,
-            detail=f"Tool '{req.name}' is not in the allowlist.",
-        )
-
-    spec = assistant_tools.TOOL_ALLOWLIST[req.name]
-    requires_conf = spec["requires_confirmation"]
-
-    if requires_conf and req.confirmation is None:
-        raise HTTPException(
-            status_code=403,
-            detail=f"Tool '{req.name}' requires explicit operator confirmation.",
-        )
-
-    # Execute via assistant_tools
-    try:
-        outcome_data = await assistant_tools.execute_tool(
-            tool_name=req.name,
-            tool_call_id=req.tool_call_id,
-            inputs=req.input,
-            confirmation=req.confirmation.model_dump() if req.confirmation else None,
-            principal=principal.id,
-            on_behalf_of=(req.confirmation.on_behalf_of or "") if req.confirmation else "",
-            correlation_id=(req.confirmation.correlation_id or "") if req.confirmation else "",
-            gh_api_fn=gh_api,
-            run_cmd_fn=run_cmd,
-            normalize_repository_fn=_normalize_repository_input,
-            org=ORG,
-            repo_root=REPO_ROOT,
-        )
-        return {
-            "success": True,
-            "tool_call_id": req.tool_call_id,
-            "name": req.name,
-            "result": outcome_data.get("result"),
-            "audit_id": outcome_data.get("audit_entry", {}).get("timestamp"),
-        }
-    except PermissionError as exc:
-        raise HTTPException(status_code=403, detail=str(exc)) from exc
-    except KeyError as exc:
-        raise HTTPException(status_code=422, detail=str(exc)) from exc
-    except Exception as exc:  # noqa: BLE001
-        log.error("tool execute error tool=%s: %s", req.name, exc)
-        # Record a failed audit entry manually if execute_tool failed before recording
-        audit_entry = assistant_tools._record_audit(  # noqa: SLF001
-            tool_name=req.name,
-            tool_call_id=req.tool_call_id,
-            inputs=req.input,
-            outcome=f"error: {exc}",
-            success=False,
-            approved_by=req.confirmation.approved_by if req.confirmation else "n/a",
-            principal=principal.id,
-            on_behalf_of=(req.confirmation.on_behalf_of or "") if req.confirmation else "",
-            correlation_id=(req.confirmation.correlation_id or "") if req.confirmation else "",
-            note=req.confirmation.note if req.confirmation else "",
-        )
-        return {
-            "success": False,
-            "tool_call_id": req.tool_call_id,
-            "name": req.name,
-            "result": {"error": str(exc)},
-            "audit_id": audit_entry["timestamp"],
-        }
-
-
-@app.get("/api/assistant/audit-history", tags=["assistant"])
-async def get_tool_audit_history(limit: int = 50) -> dict:
-    """Return the most recent assistant tool-execution audit entries (Issue #89)."""
-    capped = max(1, min(limit, 200))
-    entries = assistant_tools.get_audit_history(limit=capped)
-    return {"entries": entries, "total": len(entries)}
-
-
-# ─── Assistant Action Proposal API (Issue #89) ────────────────────────────────
-
-
-# In-memory storage for proposed actions (in real impl, use database)
-_proposed_actions: dict[str, dict] = {}
-
-
-@app.post("/api/assistant/propose-action", tags=["assistant"])
-async def propose_action(request: Request, *, principal: Principal = Depends(require_scope("assistant.chat"))) -> dict:  # noqa: B008
-    """Propose an action based on user request, awaiting operator approval."""
-    try:
-        body = await request.json()
-    except Exception:  # noqa: BLE001
-        raise HTTPException(status_code=400, detail="Invalid JSON") from None
-
-    try:
-        req = assistant_contract.ActionProposeRequest(**body)
-    except Exception as e:  # noqa: BLE001
-        raise HTTPException(status_code=422, detail=str(e)) from e
-
-    # Call AI provider to generate action proposal
-    provider_id = req.provider or "ollama_local"
-    availability = agent_remediation.probe_provider_availability()
-    if provider_id not in availability or not availability[provider_id].available:
-        # For MVP: return a synthetic proposal
-        pass  # Continue with mock response
-
-    try:
-        # MVP: return a synthetic proposal instead of calling a provider
-        response_text = json.dumps(
-            {
-                "action_type": "restart_runner",
-                "description": f"Restart runner based on request: {req.user_request[:50]}",
-                "parameters": {"runner_name": "auto"},
-                "risk_level": "medium",
-                "rationale": "This action may resolve the issue",
-                "estimated_duration_seconds": 60,
-            }
-        )
-        # Parse JSON response
-        import json as _json_parser
-
-        try:
-            proposal_dict = _json_parser.loads(response_text)
-        except Exception:  # noqa: BLE001
-            # Fallback: wrap response in a generic action
-            proposal_dict = {
-                "action_type": "custom_response",
-                "description": response_text[:200],
-                "parameters": {},
-                "risk_level": "medium",
-                "rationale": "AI-generated action",
-            }
-
-        # Generate action ID and store
-        action_id = secrets.token_hex(8)
-        _proposed_actions[action_id] = {
-            "created_at": datetime.now(UTC).isoformat(),
-            "proposal": proposal_dict,
-            "approved": False,
-        }
-
-        return {
-            "action_id": action_id,
-            "action_type": proposal_dict.get("action_type", "custom"),
-            "parameters": proposal_dict.get("parameters", {}),
-            "description": proposal_dict.get("description", ""),
-            "risk_level": proposal_dict.get("risk_level", "medium"),
-            "rationale": proposal_dict.get("rationale", ""),
-            "estimated_duration_seconds": proposal_dict.get("estimated_duration_seconds"),
-        }
-    except Exception as e:  # noqa: BLE001
-        log.error("Action proposal error: %s", e)
-        raise HTTPException(status_code=502, detail=f"AI provider error: {str(e)}") from e
-
-
-@app.post("/api/assistant/execute-action", tags=["assistant"])
-async def execute_action(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("assistant.execute")),  # noqa: B008
-) -> dict:
-    """Execute a proposed action after operator approval."""
-    try:
-        body = await request.json()
-    except Exception:  # noqa: BLE001
-        raise HTTPException(status_code=400, detail="Invalid JSON") from None
-
-    try:
-        req = assistant_contract.ActionExecuteRequest(**body)
-    except Exception as e:  # noqa: BLE001
-        raise HTTPException(status_code=422, detail=str(e)) from e
-
-    # Check action exists and is pending
-    if req.action_id not in _proposed_actions:
-        raise HTTPException(status_code=404, detail="Action not found")
-
-    action_record = _proposed_actions[req.action_id]
-    if action_record.get("approved"):
-        raise HTTPException(status_code=409, detail="Action already executed")
-
-    if not req.approved:
-        # Just mark as rejected
-        action_record["approved"] = True
-        action_record["result"] = "Rejected by operator"
-        return {
-            "success": False,
-            "action_id": req.action_id,
-            "result": "Action rejected",
-            "execution_time_ms": 0,
-        }
-
-    # Mark as approved
-    action_record["approved"] = True
-    action_record["approved_at"] = datetime.now(UTC).isoformat()
-    action_record["approved_by"] = "operator"
-    action_record["operator_notes"] = req.operator_notes
-
-    proposal = action_record["proposal"]
-    action_type = proposal.get("action_type", "unknown")
-
-    start_time = time.time()
-
-    try:
-        # Execute action based on type
-        result = "Action executed successfully"
-
-        if action_type == "restart_runner":
-            runner_name = proposal.get("parameters", {}).get("runner_name")
-            if runner_name:
-                result = f"Runner '{runner_name}' restart initiated"
-                # In real impl, would call actual restart logic
-
-        elif action_type == "rerun_workflow":
-            workflow_id = proposal.get("parameters", {}).get("workflow_id")
-            if workflow_id:
-                result = f"Workflow {workflow_id} rerun initiated"
-
-        elif action_type == "dismiss_alert":
-            alert_id = proposal.get("parameters", {}).get("alert_id")
-            if alert_id:
-                result = f"Alert {alert_id} dismissed"
-
-        execution_time_ms = int((time.time() - start_time) * 1000)
-
-        action_record["result"] = result
-        action_record["execution_time_ms"] = execution_time_ms
-
-        return {
-            "success": True,
-            "action_id": req.action_id,
-            "result": result,
-            "execution_time_ms": execution_time_ms,
-        }
-
-    except Exception as e:  # noqa: BLE001
-        execution_time_ms = int((time.time() - start_time) * 1000)
-        error_msg = str(e)
-        action_record["result"] = f"Execution failed: {error_msg}"
-        action_record["execution_time_ms"] = execution_time_ms
-        log.error("Action execution error: %s", e)
-        return {
-            "success": False,
-            "action_id": req.action_id,
-            "result": f"Execution failed: {error_msg}",
-            "execution_time_ms": execution_time_ms,
-        }
-
 
 # ─── Job Queue API ───────────────────────────────────────────────────────────
 
@@ -5732,245 +4833,6 @@ async def dispatch_assessment(
         )
     return {"status": "dispatched", "repository": repo, "provider": provider}
 
-
-# ─── Feature Requests ─────────────────────────────────────────────────────────
-
-_FEATURE_REQUESTS_PATH = Path.home() / "actions-runners" / "dashboard" / "feature_requests.json"
-_PROMPT_TEMPLATES_PATH = Path.home() / "actions-runners" / "dashboard" / "prompt_templates.json"
-_PROMPT_NOTES_PATH = Path.home() / "actions-runners" / "dashboard" / "prompt_notes.json"
-
-STANDARDS_INJECTION: dict[str, str] = {
-    "tdd": (
-        "Use Test-Driven Development: write failing tests first (RED), then minimal code to pass (GREEN),"
-        " then refactor. Tests must pass before any PR."
-    ),
-    "dbc": (
-        "Apply Design by Contract: validate inputs at boundaries, assert internal invariants,"
-        " document pre/postconditions in docstrings."
-    ),
-    "dry": (
-        "Apply DRY: extract shared logic into modules, eliminate duplication."
-        " Three similar code blocks should become one shared function."
-    ),
-    "lod": (
-        "Apply Law of Demeter: components talk to immediate neighbors only."
-        " UI receives view models, not raw nested payloads."
-    ),
-    "security": (
-        "Apply security-first: validate all inputs, avoid injection vulnerabilities,"
-        " use parameterized queries, never log secrets."
-    ),
-    "docs": (
-        "Document public APIs, non-obvious decisions, and architecture choices."
-        " Prefer short clear docstrings over multi-paragraph ones."
-    ),
-}
-
-
-@app.get("/api/feature-requests")
-async def list_feature_requests() -> dict:
-    """List saved feature implementation requests."""
-    try:
-        if _FEATURE_REQUESTS_PATH.exists():
-            data = json.loads(_FEATURE_REQUESTS_PATH.read_text(encoding="utf-8"))
-        else:
-            data = []
-    except Exception:  # noqa: BLE001
-        data = []
-    return {"requests": list(reversed(data[-100:])), "total": len(data)}
-
-
-@app.get("/api/feature-requests/templates")
-async def list_prompt_templates() -> dict:
-    """List saved prompt templates and global prompt notes."""
-    templates_data = []
-    try:
-        if _PROMPT_TEMPLATES_PATH.exists():
-            templates_data = json.loads(_PROMPT_TEMPLATES_PATH.read_text(encoding="utf-8"))
-    except Exception:  # noqa: BLE001
-        pass
-
-    prompt_notes_data = {"notes": "", "enabled": True}
-    try:
-        if _PROMPT_NOTES_PATH.exists():
-            prompt_notes_data = json.loads(_PROMPT_NOTES_PATH.read_text(encoding="utf-8"))
-    except Exception:  # noqa: BLE001
-        pass
-
-    return {
-        "templates": templates_data,
-        "standards": STANDARDS_INJECTION,
-        "promptNotes": prompt_notes_data,
-    }
-
-
-@app.post("/api/feature-requests/templates")
-async def save_prompt_template(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("feature-requests.manage")),  # noqa: B008
-) -> dict:
-    """Save a prompt template."""
-    body = await request.json()
-    name = str(body.get("name", "")).strip()
-    content = str(body.get("content", "")).strip()
-    if not name or not content:
-        raise HTTPException(status_code=422, detail="name and content required")
-    async with _prompt_templates_lock:
-        try:
-            templates: list[dict] = []
-            if _PROMPT_TEMPLATES_PATH.exists():
-                templates = json.loads(_PROMPT_TEMPLATES_PATH.read_text(encoding="utf-8"))
-            existing_idx = next((i for i, t in enumerate(templates) if t.get("name") == name), None)
-            template = {
-                "name": name,
-                "content": content,
-                "updated_at": datetime.now(UTC).isoformat(),
-            }
-            if existing_idx is not None:
-                templates[existing_idx] = template
-            else:
-                templates.append(template)
-            config_schema.atomic_write_json(_PROMPT_TEMPLATES_PATH, templates)
-        except Exception as e:  # noqa: BLE001
-            raise HTTPException(status_code=500, detail=str(e)) from e
-    return {"status": "saved", "name": name}
-
-
-@app.get("/api/settings/prompt-notes")
-async def get_prompt_notes() -> dict:
-    """Get the global prompt notes that are automatically injected into every prompt."""
-    try:
-        if _PROMPT_NOTES_PATH.exists():
-            data = json.loads(_PROMPT_NOTES_PATH.read_text(encoding="utf-8"))
-        else:
-            data = {"notes": "", "enabled": True}
-    except Exception:  # noqa: BLE001
-        data = {"notes": "", "enabled": True}
-    return data
-
-
-@app.put("/api/settings/prompt-notes")
-async def update_prompt_notes(request: Request, *, principal: Principal = Depends(require_scope("operator"))) -> dict:  # noqa: B008
-    """Update the global prompt notes."""
-    body = await request.json()
-    if not isinstance(body, dict):
-        raise HTTPException(status_code=422, detail="expected object body")
-
-    notes = str(body.get("notes", "")).strip()
-    enabled = bool(body.get("enabled", True))
-
-    async with _prompt_notes_lock:
-        try:
-            data = {"notes": notes, "enabled": enabled}
-            config_schema.atomic_write_json(_PROMPT_NOTES_PATH, data)
-        except Exception as e:  # noqa: BLE001
-            raise HTTPException(status_code=500, detail=str(e)) from e
-    return {"status": "saved", "notes_length": len(notes), "enabled": enabled}
-
-
-@app.post("/api/feature-requests/dispatch")
-async def dispatch_feature_request(
-    request: Request,
-    *,
-    principal: Principal = Depends(require_scope("feature-requests.manage")),  # noqa: B008
-) -> dict:
-    """Dispatch a feature implementation request via CI remediation workflow."""
-    client_ip = request.client.host if request.client else "unknown"
-    check_dispatch_rate(client_ip)
-    body = await request.json()
-    repo = str(body.get("repository", "")).strip()
-    branch = str(body.get("branch", "main")).strip()
-    provider = str(body.get("provider", "jules_api")).strip()
-    prompt = str(body.get("prompt", "")).strip()
-    standards = body.get("standards", []) or []
-    template_id = str(body.get("template_id", "")).strip()
-    if not repo:
-        raise HTTPException(status_code=422, detail="repository required")
-    if not prompt and not template_id:
-        raise HTTPException(status_code=422, detail="prompt or template_id required")
-
-    log.info(
-        "audit: feature_request_dispatch repo=%s provider=%s branch=%s",
-        sanitize_log_value(repo),
-        sanitize_log_value(provider),
-        sanitize_log_value(branch),
-    )
-
-    # Load and apply prompt notes if enabled
-    prompt_notes_data: dict[str, object] = {"notes": "", "enabled": True}
-    try:
-        if _PROMPT_NOTES_PATH.exists():
-            prompt_notes_data = json.loads(_PROMPT_NOTES_PATH.read_text(encoding="utf-8"))
-    except Exception:  # noqa: BLE001
-        pass
-
-    # Build full prompt with notes and standards injection
-    full_prompt = prompt
-    notes_val = str(prompt_notes_data.get("notes", ""))
-    if prompt_notes_data.get("enabled", True) and notes_val.strip():
-        full_prompt = f"{notes_val}\n\n{prompt}"
-
-    injected_standards = "\n\n".join(
-        f"[{s.upper()}] {STANDARDS_INJECTION[s]}" for s in standards if s in STANDARDS_INJECTION
-    )
-    if injected_standards:
-        full_prompt = f"{full_prompt}\n\n## Engineering Standards\n{injected_standards}"
-
-    # Save to history
-    entry: dict = {}
-    async with _feature_requests_lock:
-        try:
-            history: list[dict] = []
-            if _FEATURE_REQUESTS_PATH.exists():
-                history = json.loads(_FEATURE_REQUESTS_PATH.read_text(encoding="utf-8"))
-            entry = {
-                "id": str(int(datetime.now(UTC).timestamp())),
-                "repository": repo,
-                "branch": branch,
-                "provider": provider,
-                "prompt": prompt[:500],
-                "standards": list(standards),
-                "status": "dispatched",
-                "created_at": datetime.now(UTC).isoformat(),
-            }
-            history.append(entry)
-            config_schema.atomic_write_json(_FEATURE_REQUESTS_PATH, history[-200:])
-        except Exception:  # noqa: BLE001
-            pass
-
-    # Dispatch via feature-request workflow
-    endpoint = f"/repos/{ORG}/Repository_Management/actions/workflows/Jules-Feature-Request.yml/dispatches"
-    payload = {
-        "ref": "main",
-        "inputs": {
-            "target_repository": f"{ORG}/{repo}",
-            "branch": branch,
-            "provider": provider,
-            "prompt": full_prompt[:10000],
-        },
-    }
-    with tempfile.NamedTemporaryFile(mode="w", encoding="utf-8", suffix=".json", delete=False) as f:
-        json.dump(payload, f)
-        pf = f.name
-    try:
-        code, _, stderr = await run_cmd(
-            ["gh", "api", endpoint, "--method", "POST", "--input", pf],
-            timeout=30,
-            cwd=REPO_ROOT,
-        )
-    finally:
-        with contextlib.suppress(OSError):
-            Path(pf).unlink()
-    if code != 0:
-        log.warning("feature_request_dispatch failed: %s", stderr.strip()[:200])
-        # Don't raise - save history record and return success anyway (workflow may not exist yet)
-    return {
-        "status": "dispatched",
-        "repository": repo,
-        "provider": provider,
-        "entry_id": entry.get("id", ""),
-    }
 
 
 # ─── Fleet Orchestration Control Plane ───────────────────────────────────────
